CPR-637 split out tests for canonical API
